# docs/swagger.yaml
basePath: /
definitions:
  model.Conversion:
    properties:
      amount:
        type: number
      from:
        $ref: '#/definitions/model.Currency'
      result:
        type: number
      to:
        $ref: '#/definitions/model.Currency'
    type: object
  model.ConversionRequest:
    properties:
      amount:
        type: number
      from:
        type: string
      to:
        type: string
    type: object
  model.Currency:
    properties:
      code:
        type: string
      name:
        type: string
      rate:
        type: number
      symbol:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления валютами и конвертациями
  title: Currency Converter API
  version: "1.0"
paths:
  /conversion:
    post:
      consumes:
      - application/json
      description: Converts amount from one currency to another using current Central
        Bank of Russia exchange rates and saves the conversion result
      parameters:
      - description: Conversion request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ConversionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully converted currency
          schema:
            $ref: '#/definitions/model.Conversion'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Currency not found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Invalid conversion parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Convert currency amount
      tags:
      - conversion
  /conversions:
    get:
      description: Retrieves history of all currency conversions performed
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved conversion history
          schema:
            items:
              $ref: '#/definitions/model.Conversion'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get conversion history
      tags:
      - conversion
  /currencies:
    get:
      description: Retrieves all currencies with current exchange rates from Central
        Bank of Russia
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved currencies map
          schema:
            additionalProperties:
              $ref: '#/definitions/model.Currency'
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get list of all available currencies
      tags:
      - currency
  /currency:
    post:
      consumes:
      - application/json
      description: Adds a new currency to the storage
      parameters:
      - description: Currency data
        in: body
        name: currency
        required: true
        schema:
          $ref: '#/definitions/model.Currency'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Currency'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create currency
      tags:
      - currency
  /currency/{code}:
    get:
      description: Retrieves detailed information about specific currency including
        exchange rate from Central Bank of Russia
      parameters:
      - description: Currency code (ISO 4217 format)
        example: USD
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved currency details
          schema:
            $ref: '#/definitions/model.Currency'
        "400":
          description: Invalid currency code format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Currency not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get currency details by code
      tags:
      - currency
    put:
      consumes:
      - application/json
      description: 'Updates the exchange rate for a specific currency. Note: Normally
        rates are updated automatically from Central Bank of Russia'
      parameters:
      - description: Currency code to update (ISO 4217 format)
        example: USD
        in: path
        name: code
        required: true
        type: string
      - description: Currency data with updated exchange rate
        in: body
        name: currency
        required: true
        schema:
          $ref: '#/definitions/model.Currency'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated currency
          schema:
            $ref: '#/definitions/model.Currency'
        "400":
          description: Invalid input data or currency code mismatch
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Currency not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update currency exchange rate
      tags:
      - currency
swagger: "2.0"